introducing single_losses array in schema
because mu, median loss and sigma are all based on a list of single loss amounts, it makes sense for the standard to support this "root data" and let the engines calculate median loss, mu and sigma internally. Also I fixed a bug with currency exchange. Still some tests are failing. Looking into it.